relay_with_PIR.ino added
Connect the PIR motion sensor's Vcc and GND to the 7805's output. Connect the BC547 (pin out - Collector, Base, Emitter) collector to one end of the relay coil. Connect the base to the PIR motion sensor's output signal and the emitter to theGNDd. The other end of the relay's coil is connected to the input voltage (9V).22-Mar-2018